requirements.txt
numpy
scipy
matplotlib
cdo
netcdf4
xarray
pyyaml
cdo
numpy
netcdf4
matplotlib
pyyaml
scipy
xarray